pyproject.toml
[project]
name = "function-health"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "assertpy>=1.1",
    "black>=26.1.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pylint>=4.0.4",
    "python-dotenv>=1.0.1",
    "pytest>=9.0.2",
    "pytest-playwright>=0.7.2",
]
